Create App folder and move App component and its CSS file there
